% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab.R
\name{ab}
\alias{ab}
\alias{ab.cdf}
\alias{ab.pdf}
\alias{ab.orig}
\alias{ab.inverse}
\alias{ab.orig.cdf}
\alias{ab.orig.pdf_x}
\alias{ab.inverse_x.cdf}
\alias{ab.inverse_x.pdf_x}
\title{ab}
\usage{
ab(x, params)

ab.orig(x, params)

ab.inverse(x, params)

\method{ab}{orig.cdf}(x, params)

\method{ab}{orig.pdf_x}(x, params)

\method{ab}{inverse_x.cdf}(x, params)

\method{ab}{inverse_x.pdf_x}(x, params)
}
\arguments{
\item{x}{Vector of x parametres a and b}

\item{params}{Function ab has two parameters: \n a - linear parameter \n b nominater parameter}
}
\value{
Vector of result vaues
}
\description{
AB function is a core type function.
\cr
\cr
\cr
It's CDF formula is: y = (x - a) / b
\cr
It's PDF formula per x is: y = 1/b
\cr
It's PDF formula per a is: y = -1/b
\cr
It's PDF formula per b is: y = (x - a) / b^2
}
