% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAdjusted.R
\name{al}
\alias{al}
\alias{al.cdf}
\alias{al.pdf}
\alias{al.orig}
\alias{al.inverse}
\alias{al.orig.cdf}
\alias{al.orig.pdf_x}
\alias{al.inverse_x.cdf}
\alias{al.inverse_x.pdf_x}
\title{Linear adjusted function}
\usage{
al(x, params)

al.orig(x, params)

al.inverse(x, params)

\method{al}{orig.cdf}(x, params)

\method{al}{orig.pdf_x}(x, params)

\method{al}{inverse_x.cdf}(x, params)

\method{al}{inverse_x.pdf_x}(x, params)
}
\arguments{
\item{x}{Vector of x parametres a and b}

\item{params}{Function la has two parameters: \n a - threshold parameter \n b slope parameter}
}
\value{
Vector of result vaues
}
\description{
AL function is a core type function.
\cr
\cr
\cr
It's CDF formula is: y = (x - a) * b
\cr
It's PDF formula is: y per x= b
\cr
It's PDF formula per a is: y = -b
\cr
It's PDF formula per b is: y = (x-a)
}
