% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy.R
\name{cauchy}
\alias{cauchy}
\alias{cauchy.cdf}
\alias{cauchy.pdf}
\alias{cauchy.orig}
\alias{cauchy.inverse}
\alias{cauchy.orig.cdf}
\alias{cauchy.orig.pdf}
\alias{cauchy.inverse.cdf}
\alias{cauchy.inverse.pdf}
\title{Cauchy}
\usage{
cauchy(x)

cauchy.orig(x)

cauchy.inverse(x)

\method{cauchy}{orig.cdf}(x)

\method{cauchy}{orig.pdf}(x)

\method{cauchy}{inverse.cdf}(x)

\method{cauchy}{inverse.pdf}(x)
}
\arguments{
\item{x}{Vector of x parametres}
}
\value{
Vector of result vaues
}
\description{
Cauchy function is a sigmoid type function.
\cr
It's CDF formula is: y = atan(x) / pi + 0.5
\cr
It's PDF formula is: y = 1/ (1 + x^2) / pi
}
