% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.R
\name{def}
\alias{def}
\title{def}
\usage{
def(data, sigmoid, core, par = NULL, fn = NULL, gr = NULL, ...,
  method = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent"),
  lower = -Inf, upper = Inf, control = list(), hessian = FALSE)
}
\arguments{
\item{data}{Specifies the data set on which the function will be fitted.  Data have to be formated in specified way - data.frame/tibble (yes, no, predictor columns).}

\item{sigmoid}{determines the outer shape of the fuction}

\item{core}{dermines scalling of predictor}

\item{...}{specifies the parametres of optim function}

\item{gamma}{sets the loves boundary of function}
}
\value{
vector of return values
}
\description{
Default algorithm for fitting parameters of psychometric function. Version of algorithm with fixed gamma parameter. Gamma has to be specifed.
}
