% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_fixed_gamma_lambda.R
\name{def_fixed_gamma_lambda}
\alias{def_fixed_gamma_lambda}
\title{def}
\usage{
def_fixed_gamma_lambda(data, sigmoid, core, gamma = 0.05,
  lambda = 0.05, par = NULL, fn = NULL, gr = NULL, ...,
  method = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent"),
  lower = -Inf, upper = Inf, control = list(), hessian = FALSE)
}
\arguments{
\item{data}{Specifies the data set on which the function will be fitted.  Data have to be formated in specified way - data.frame/tibble (yes, no, predictor columns).}

\item{sigmoid}{determines the outer shape of the fuction}

\item{core}{dermines scalling of predictor}

\item{gamma}{sets the loves boundary of function}

\item{lambda}{sets the highes boundary of function}

\item{...}{specifies the parametres of optim function}
}
\value{
vector of return values
}
\description{
Default algorithm for fitting parameters of psychometric function. Only inner parameters are fitted, gamma and lambda are fixed and have to be specified.
}
