% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprime_threshold.R
\name{dprime_threshold}
\alias{dprime_threshold}
\title{D-prime based Threshold}
\usage{
dprime_threshold(pf, reference = 0, d = 1, transoform = NULL)
}
\arguments{
\item{pf}{Psychometric function model}

\item{reference}{Stimulus level used as base for measuring d-prime distance}

\item{d}{D-prime distance between two stimuli}
}
\value{
Level of stimuli that is in d distance from point of reference
}
\description{
Estimates threshold level of given function based on given paramters.
\cr
Improvement based threshold is a point where subject's performance grows the most.
}
