% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponetial.R
\name{exponetial}
\alias{exponetial}
\alias{exponential}
\alias{exponetial.cdf}
\alias{exponetial.pdf}
\title{Exponetial function}
\usage{
exponential(x)
}
\arguments{
\item{x}{Vector of x parametres}
}
\value{
Vector of result vaues
}
\description{
Exponential function is a sigmoid type function.
\cr
It's CDF formula is: y = 1 - exp(-x))
\cr
It's PDF formula is: y = exp(-x))
}
