% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPFm.R
\name{fitPFm}
\alias{fitPFm}
\title{fitPFm}
\usage{
fitPFm(formula, data, sigmoid, core, ..., split_by = NULL,
  type = "yes/no", algorithm = def)
}
\arguments{
\item{formula}{Formula to identify important componets of data. Form is c(yes,no)~predictor for yes/no notation and c(PC,Observations)~predictor for Proportion correct notation.}

\item{sigmoid}{determines the sigmoid of the fuction}

\item{core}{dermines the core of the function}

\item{...}{specifies the parametres for used fitting algorithm}

\item{split_by}{specifies, whether and by which criteria should data be divided into subgroups (observers, sessions)}

\item{type}{specifies type of data formating, proportion correct or yes/no}

\item{algrithm}{specifies the algorithm to be used for fitting psychometric function parameters.}
}
\value{
vector of return values
}
\description{
Estimates parameters of psychometric function based on given data and settings with given algorithm.
}
