% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss.R
\name{gauss}
\alias{gauss}
\alias{gauss.cdf}
\alias{gauss.pdf}
\alias{gauss.orig}
\alias{gauss.inverse}
\alias{gauss.orig.cdf}
\alias{gauss.orig.pdf}
\alias{gauss.inverse.cdf}
\alias{gauss.inverse.pdf}
\title{Gauss}
\usage{
gauss(x)

gauss.orig(x)

gauss.inverse(x)

\method{gauss}{orig.cdf}(x)

\method{gauss}{orig.pdf}(x)

\method{gauss}{inverse.cdf}(x)

\method{gauss}{inverse.pdf}(x)
}
\arguments{
\item{x}{Vector of x parametres}
}
\value{
Vector of result vaues
}
\description{
Gauss function is a sigmoid type function.
\cr
It's CDF formula is: y = pnorm(x,mean = 0, sd=1, log=FALSE)
\cr
It's PDF formula is: y = exp(-x^2/2) / (2*pi)^(1/2)
}
