% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel_l.R
\name{gumbel_l}
\alias{gumbel_l}
\alias{gumbel_l.cdf}
\alias{gumbel_l.pdf}
\alias{gumbel_l.orig}
\alias{gumbel_l.inverse}
\alias{gumbel_l.orig.cdf}
\alias{gumbel_l.orig.pdf}
\alias{gumbel_l.inverse.cdf}
\alias{gumbel_l.inverse.pdf}
\title{Gumbel_l}
\usage{
gumbel_l(x)

gumbel_l.orig(x)

gumbel_l.inverse(x)

\method{gumbel_l}{orig.cdf}(x)

\method{gumbel_l}{orig.pdf}(x)

\method{gumbel_l}{inverse.cdf}(x)

\method{gumbel_l}{inverse.pdf}(x)
}
\arguments{
\item{x}{Vector of x parametres}
}
\value{
Vector of result vaues
}
\description{
Left gumbel function is a sigmoid type function.
\cr
It's CDF formula is: y = 1 - exp(- exp(x))
\cr
It's PDF formula is: y = exp(- exp(x) + x)
}
