% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel_r.R
\name{gumbel_r}
\alias{gumbel_r}
\alias{gumbel_r.orig}
\alias{gumbel_r.inverse}
\alias{gumbel_r.orig.cdf}
\alias{gumbel_r.orig.pdf}
\alias{gumbel_r.inverse.cdf}
\alias{gumbel_r.inverse.pdf}
\title{Gumbel_r}
\usage{
gumbel_r(x)

gumbel_r.orig(x)

gumbel_r.inverse(x)

\method{gumbel_r}{orig.cdf}(x)

\method{gumbel_r}{orig.pdf}(x)

\method{gumbel_r}{inverse.cdf}(x)

\method{gumbel_r}{inverse.pdf}(x)
}
\arguments{
\item{x}{Vector of x parametres}
}
\value{
Vector of result vaues
}
\description{
Right gumbel function is a sigmoid type function.
\cr
It's CDF formula is: exp(- exp(-x))
\cr
It's PDF formula is: y = exp(- exp(-x) - x)
}
