% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heu_fixed_gamma.R
\name{heu_fixed_gamma}
\alias{heu_fixed_gamma}
\title{heu}
\usage{
heu_fixed_gamma(data, sigmoid, core, gamma = 0.05, par = NULL,
  fn = NULL, gr = NULL, ..., method = c("Nelder-Mead", "BFGS", "CG",
  "L-BFGS-B", "SANN", "Brent"), lower = -Inf, upper = Inf,
  control = list(), hessian = FALSE)
}
\arguments{
\item{data}{Specifies the data set on which the function will be fitted.  Data have to be formated in specified way - data.frame/tibble (yes, no, predictor columns).}

\item{sigmoid}{determines the outer shape of the fuction}

\item{core}{dermines scalling of predictor}

\item{gamma}{sets the loves boundary of function}

\item{...}{specifies the parametres of optim function}
}
\value{
vector of return values
}
\description{
Heuristically improved algorithm for fitting parameters of psychometric function. Version of algorithm with fixed gamma parameter. Gamma has to be specifed.
}
