% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htan.R
\name{htan}
\alias{htan}
\alias{htan.cdf}
\alias{htan.pdf}
\alias{htan.orig}
\alias{htan.inverse}
\alias{htan.orig.cdf}
\alias{htan.orig.pdf}
\alias{htan.inverse.cdf}
\alias{htan.inverse.pdf}
\title{Exponetial function}
\usage{
htan(x)

htan.orig(x)

htan.inverse(x)

\method{htan}{orig.cdf}(x)

\method{htan}{orig.pdf}(x)

\method{htan}{inverse.cdf}(x)

\method{htan}{inverse.pdf}(x)
}
\arguments{
\item{x}{Vector of x parametres}
}
\value{
Vector of result vaues
}
\description{
htan function is a sigmoid type function.
\cr
It's CDF formula is: y = 0.5 + tanh(x)/2)
\cr
It's PDF formula is: y = 0.5/(cosh(x)^2)
}
