% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_threshold.R
\name{imp_threshold}
\alias{imp_threshold}
\title{Improvement based Threshold}
\usage{
imp_threshold(pf)
}
\arguments{
\item{pf}{hit percentage of experiment data}
}
\value{
Level of stimuli with highest derivative
}
\description{
Estimates threshold level of given function based on given paramters.
\cr
Improvement based threshold is a point where subject's performance grows the most.
}
