% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR.R
\name{iqr}
\alias{iqr}
\title{Interquartile range}
\usage{
iqr(pf)
}
\arguments{
\item{pf}{Psychometric function representation the IQR is to be estimated}
}
\value{
The distance beween lower and upper quartiles of inner function
}
\description{
Estimates interquartile range distance of given function based on given paramters.
\cr
Interquartile range is a distance between 1/4 and 3/4 'up' of inner psychometric function
}
