% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{linear}
\alias{linear}
\alias{linear.cdf}
\alias{linear.pdf}
\alias{linear.orig}
\alias{linear.inverse}
\alias{linear.orig.cdf}
\alias{linear.orig.pdf_x}
\alias{linear.orig.pdf_p1}
\alias{linear.orig.pdf_p2}
\alias{linear.inverse_x.cdf}
\alias{linear.inverse_x.pdf_x}
\title{Linear}
\usage{
linear(x, params)

linear.orig(x, params)

linear.inverse(x, params)

\method{linear}{orig.cdf}(x, params)

\method{linear}{orig.pdf_x}(x, params)

\method{linear}{orig.pdf_p1}(x, params)

\method{linear}{orig.pdf_p2}(x, params)

\method{linear}{inverse_x.cdf}(x, params)

\method{linear}{inverse_x.pdf_x}(x, params)
}
\arguments{
\item{x}{Vector of x parametres a and b}

\item{params}{Linear function has to paramters: \n a - relative paramter \n b - absolute paramter}
}
\value{
Vector of result vaues
}
\description{
Linear function is a core type function.
\cr
\cr
\cr
It's CDF formula is: y = a * x + b
\cr
It's PDF formula per x is: y = a
\cr
It's PDF formula per a is: y = x
\cr
It's PDF formula per b is: y = 1
}
