% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Log-Likelihood of psychometric model}
\usage{
log_likelihood(pfm)
}
\arguments{
\item{pfm}{Psychometric function model}
}
\value{
Likelihood value of model
}
\description{
Estimates likelihood of given model based on given data prediction.
\cr
}
