% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likeliRatio.R
\name{log_likRatio}
\alias{log_likRatio}
\title{Log-Likelihood deviance of psychometric model}
\usage{
log_likRatio(pfm)
}
\arguments{
\item{pfm}{Psychometric function model}
}
\value{
Value of likelihood ratio
}
\description{
Estimates likelihood ratio/deviance of given model and ideal model based on given data.
}
