% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\alias{logistic.cdf}
\alias{logistic.pdf}
\alias{logistic.orig}
\alias{logistic.inverse}
\alias{logistic.orig.cdf}
\alias{logistic.orig.pdf}
\alias{logistic.inverse.cdf}
\alias{logistic.inverse.pdf}
\title{Logistic}
\usage{
logistic(x)

logistic.orig(x)

logistic.inverse(x)

\method{logistic}{orig.cdf}(x)

\method{logistic}{orig.pdf}(x)

\method{logistic}{inverse.cdf}(x)

\method{logistic}{inverse.pdf}(x)
}
\arguments{
\item{x}{Vector of x parametres}
}
\value{
Vector of result vaues
}
\description{
Logistic function is a sigmoid type function. It is the part of CDF.
\cr
It's CDF formula is: y = 1 / ( 1 + exp(-x))
\cr
It's PDF formula is: y = exp(x) / ( 1 + exp(x))^2
}
