% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{MSE psychometric model}
\usage{
mse(pfm)
}
\arguments{
\item{pfm}{Psychometric function model}
}
\value{
MSE measuremet of model
}
\description{
Estimates MSE of given model based on given data.
}
