% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisedPredict.R
\name{noisedPredict}
\alias{noisedPredict}
\title{noisedPredict}
\usage{
noisedPredict(pfm, x, noiseDistr = c("normal", "poisson", "exponential",
  "logistic", "uniform", "binomial"), ..., symetric = FALSE)
}
\arguments{
\item{pfm}{Psychometric function model}

\item{x}{the vector of level values}
}
\value{
vector of model values
}
\description{
Predicts the values of given model of psychometric function based on level of stimulation it also ads a noise to every simultaion.
}
