% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC2YesNo.R
\name{PC2YesNo}
\alias{PC2YesNo}
\title{fitPF}
\usage{
PC2YesNo(predictor, observation, PC)
}
\arguments{
\item{predictor}{the vector of predictor values}

\item{PC}{proportion of hits over all responses}

\item{observations}{vector of number of observations per predictor}

\item{inverse}{specifies, whether to compute the inverse function}
}
\value{
dd data.frame containing predictor and number of yes/no respones
}
\description{
Provides the frame for psycometrics functions. Combines the sigmoid and core function.
}
