% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearsonx.R
\name{pearsonx}
\alias{pearsonx}
\title{PearsonX2 deviance of psychometric model}
\usage{
pearsonx(pfm)
}
\arguments{
\item{pfm}{Psychometric function model}
}
\value{
PearsonX measuremet of model
}
\description{
Estimates Pearsons X^2 deviance of given model based on given data.
}
