% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_threshold.R
\name{perf_threshold}
\alias{perf_threshold}
\title{Performance based Threshold}
\usage{
perf_threshold(pf)
}
\arguments{
\item{pf}{Psychometric function representation the threshold is to be estimated}
}
\value{
Level of stimuli reaching tha halfway of inner function
}
\description{
Estimates threshold level of given function based on given paramters.
\cr
Performance based threshold is a point where subjects performance is halfway between gamma and 1-lambda.
}
