% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PF.R
\name{PF}
\alias{PF}
\title{PFunction}
\usage{
PF(sigmoid, core, gamma, lambda, params)
}
\arguments{
\item{sigmoid}{determines the sigmoid of the fuction}

\item{core}{dermines the core of the function}

\item{lambda}{sets the highes boundary of function}

\item{params}{specifies inner parametres of core function}

\item{gama}{sets the loves boundary of function}
}
\value{
vector of return values
}
\description{
Provides the frame for psycometrics functions. Combines the sigmoid and core function.
}
