% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFm.R
\name{PFm}
\alias{PFm}
\title{PFm}
\usage{
PFm(pf, predictor, observations, otherData = NULL, type = c("PC",
  "yes/no"), noise = FALSE, ...)
}
\arguments{
\item{pf}{psychometric function description of type PF}

\item{predictor}{Value of stimulus intensity, that should predict ability to distinguish stimulus}

\item{otherData}{List of data.frame with other data to be included along the generated data of PFm}

\item{type}{Type of response encoding PC, or yes/no}

\item{noise}{Should be noise included in level generation}

\item{...}{Parameters of noise generation}

\item{observation}{Number of observations over one stimulus intensity}
}
\value{
PFm model
}
\description{
Creates psychometric function model from psychometric function description
}
