% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFunction.R
\name{PFunction}
\alias{PFunction}
\title{PFunction}
\usage{
PFunction(sigmoid, core, x, gamma, lambda, ..., type = "cdf",
  inverse = FALSE, corePdf = NULL)
}
\arguments{
\item{sigmoid}{determines the sigmoid of the fuction}

\item{core}{dermines the core of the function}

\item{x}{the vector of level values}

\item{lambda}{sets the highes boundary of function}

\item{...}{specifies the parametres of core function}

\item{type}{specifies, whether function is CDF of PDF type}

\item{inverse}{specifies, whether to compute inversion of specified function}

\item{gama}{sets the loves boundary of function}
}
\value{
vector of return values
}
\description{
Function used to utilize all versions of psychometric function . Combines the sigmoid and core function.
}
