% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.PF}
\alias{predict.PF}
\title{predict}
\usage{
\method{predict}{PF}(object, x = c(), ...)
}
\arguments{
\item{object}{PF, or PFm to predict results with}

\item{x}{the vector of predictor (stimulus level) values}
}
\value{
vector of model values
}
\description{
Predicts the values of given model of psychometric function based on level of stimulation.
}
