% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primalParamsDef.R
\name{primalParamsDef}
\alias{primalParamsDef}
\title{Primal parameters}
\usage{
primalParamsDef(sigmoidi, corei_x, data)
}
\arguments{
\item{sigmoidi}{inverse of psychometric function sigmoid}

\item{corei_x}{inverse for psychometric function core}

\item{data}{data for derermining the outer shape of the fuction}
}
\description{
Default algorithm for adjusting initial parameters of core of psychometric psychometric function.
}
