% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.R
\name{quick}
\alias{quick}
\alias{quick.cdf}
\alias{quick.pdf}
\alias{quick.orig}
\alias{quick.inverse}
\alias{quick.orig.cdf}
\alias{quick.orig.pdf}
\alias{quick.inverse.cdf}
\alias{quick.inverse.pdf}
\title{Quick function}
\usage{
quick(x)

quick.orig(x)

quick.inverse(x)

\method{quick}{orig.cdf}(x)

\method{quick}{orig.pdf}(x)

\method{quick}{inverse.cdf}(x)

\method{quick}{inverse.pdf}(x)
}
\arguments{
\item{x}{Vector of x parametres}
}
\value{
Vector of result vaues
}
\description{
Quick function is a sigmoid type function.
\cr
It's CDF formula is: y = 1 - 2^(-x))
\cr
It's PDF formula is: y = log(2)*2^(-x))
}
