% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Width.R
\name{w_range}
\alias{w_range}
\title{Width}
\usage{
w_range(pf)
}
\arguments{
\item{pf}{Psychometric function representation the IQR is to be estimated}
}
\value{
The distance beween lower and upper quartiles of inner function
}
\description{
Estimates Width - range distance between quartile 0.95 and 0.05 of given function based on given paramters.
\cr
Interquartile range is a distance between 1/4 and 3/4 'up' of inner psychometric function
}
