% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YesNo2PC.R
\name{YesNo2PC}
\alias{YesNo2PC}
\title{YesNo2PC}
\usage{
YesNo2PC(predictor, yes, no)
}
\arguments{
\item{predictor}{the vector of predictor values}

\item{yes}{number yes answers per predictor value}

\item{no}{number no answers per predictor value}
}
\value{
dd data.frame containing predictor and number of yes/no respones
}
\description{
Converts data of psychometric function in yes/no format to PC format.
}
